Update installation of 'store_product' page type
When the 'store_product' page type was installed, the publish target for it wasn't set. This caused an error when adding new pages to the sitemap. This updates the published target when the 'store_product' page type gets installed. It does not, however, update the publish target if the 'store_product' page has already been installed.